Add initial window positioning proof of concept
The overall plan with the new .NET GitWatcher is to create a borderless
window that automatically positions itself to the right of the Terminal
which is running in Quake Mode. The width of the Terminal will be
adjusted by GitWatcher so that the Quake Terminal will take up the left
half of the screen and the GitWatcher window will take up the right half
of the screen.

This commit roughly implements this idea as a proof of concept. Further
refactoring should be done to make this implementation sustainable.